Fixed a variable scope problem
- Fixed a problem when the variable where assigned it wasn't vsible inside deeper scopes.

Known issue: all variables are kind of "globals" during the compilation. This should not be a problem anyway because it's at runtime thatwe'll take care of the scopes and everything. In fact, this header thing is used to retreive variables names, not to know which variable name is defined or not.

Signed-off-by: Gabriel Santamaria <gaby.santamaria@outlook.fr>